Переведены здоровье и аптечка на float

+++ Assets/Scripts/Pickable/Medicine/Medicine.cs
@@ -5,11 +5,11 @@ namespace Pickable.Medicine
 {
     public class Medicine : MonoBehaviour, IPickable
     {
-        [SerializeField, Min(0)] private int _value;
+        [SerializeField, Min(0)] private float _value;
 
         public event Action<IPickable> Picked;
 
-        public int Value =>
+        public float Value =>
             _value;
 
         public void PickUp() =>
